# Improve performance of crc32_acle on 32-bit ARM