This does a few things:
- `__crc32d` is called even on AArch32, where it expands to two `__crc32w` instructions.
- ~The number of intermediate variables is reduced, which makes modern GCC more likely to use `ldrd` or `ldm` instead of two `ldr` instructions.~ This wasn't as beneficial as I'd hoped.
- ~`size_t` is used for the length to avoid 64-bit comparisons and subtractions in the inner loop. This can be simplified once PR #1356 is merged.~
- The alignment checks are simplified so that most of them are skipped if the buffer is already aligned.
- A special case is added when the length is 1.


Results from `benchmark_zlib` on a Raspberry Pi 400 (outdated) - Before:
```
----------------------------------------------------------------------------------
Benchmark                                        Time             CPU   Iterations
----------------------------------------------------------------------------------
crc32/acle/1                                  6.22 ns         6.22 ns    112505190
crc32/acle/8                                  25.9 ns         25.9 ns     27547565
crc32/acle/64                                 37.8 ns         37.8 ns     18518587
crc32/acle/512                                 235 ns          235 ns      2977104
crc32/acle/4096                               1730 ns         1729 ns       404738
crc32/acle/32768                             13968 ns        13967 ns        50177
crc32/acle/262144                           110440 ns       110436 ns         6327
crc32/acle/2097152                          984960 ns       984810 ns          707
crc32/acle/4194304                         1966942 ns      1966880 ns          355
```

After:
```
----------------------------------------------------------------------------------
Benchmark                                        Time             CPU   Iterations
----------------------------------------------------------------------------------
crc32/acle/1                                  11.7 ns         11.7 ns     59957221
crc32/acle/8                                  16.7 ns         16.7 ns     41973016
crc32/acle/64                                 27.8 ns         27.8 ns     25184196
crc32/acle/512                                99.0 ns         98.9 ns      7074500
crc32/acle/4096                                597 ns          597 ns      1171412
crc32/acle/32768                              5081 ns         5080 ns       134381
crc32/acle/262144                            38645 ns        38641 ns        18111
crc32/acle/2097152                          559192 ns       559120 ns         1238
crc32/acle/4194304                         1118589 ns      1118471 ns          623
```