@mtl1979 2023-01-11
This seems to remove early exit for short lengths and that will cause net performance loss with short lengths due to added cost of while loop.