@mtl1979 2023-01-11
@ccawley2011 Somewhere disappears over 5.4 ns of performance and that is way more than error margin...